CHANNELMGR-2080 show message when component is deleted during editing
If another user has deleted the component that is being edited, the redraw action fails. The backend can no longer provide updated markup for the values being edited and returns a 404. In this case further editing makes no sense, to a toast message is shown, the editor is closed and the page is reloaded.