CHANNELMGR-2093 Also redraw component with invalid fields
New insights: it is better to try to redraw the component anyway after a
property change, although the form is not yet valid.

The fact that a zero is returned by the backend for an invalid number in
the mean time is OK. Forms with invalid fields should not be saved, but
redraw is allowed. The old dialog also redraws components when the form
has required fields that are invalid. So this way the behavior stays the
same.