CHANNELMGR-2118 Use defaultValue if value is empty
The legacy behavior is to use the defaultValue if it is not empty and
the actual value is empty. This means that fields with a non-empty
defaultValue cannot contain an empty string; the HST will fallback
to using the defaultValue in this case.

We decided to divert a little bit from the legacy behavior here with
regards to the UX: in the old component properties dialog, if a user
empties a field, it will be filled with the defaultValue after the dialog
has been saved. In the new situation, when the user empties a field, it
will be filled in with the defaultValue right away. This seems more in
line with what is happening, i.e., the component on the page is also
redrawn with the defaultValue and saving the empty field would result
in saving the defaultValue.