CMS-11492 encode node links in the repository servlet
Removed html and url builtins: they must be configured with a ? instead of a !, which is an expression of the default value operator. To encode urls the url_path builtin is used. It encodes for example square brackets [] but not forward slashes.

(cherry picked from commit 9bba61e1c7a75a187dd0707d1a4df72bd009d94a)