CHANNELMGR-2118 Add defaultDisplayValue for linkpicker fields
If the property value equals the defaultValue, we know that the
displayValue is the correct node name (loaded by the backend)
and store the displayValue as the defaultDisplayValue.
If the property value is different from defaultValue and defaultValue
is not empty, we try to find the last path segment and store that
as the defaultDisplayValue.

Refactored tests.