CHANNELMGR-2119 Introduce PickerService to open picker dialogs
The PickerService hides sending events to ExtJs and handling the result.
Its functions return promises, so the calling code can become a lot more
familiar.

The 'show-path-picker' event now also passed two callbacks (like
'show-link-picker' and 'show-image-picker' already did) which simplies
the handling of the results.

The rich-text based pickers are not opened via the PickerService yet.
When wrapping the callbacks in a promise the interaction with CKEditor
led to digest problems (e.g. "digest already in progress") so handling
those events is left as-is for now.