REPO-2145 Let versionRestoreTo unpublished variant
DocumentWorkflow#versionRestoreTo should return the unpublished variant,
because in the cms (org.hippoecm.addon.workflow.StdWorkflow#execute)
this is expected ( CMS-11373 ).

If a version node is returned then a new editor service is created for a
frozen node and the cluster config is updated. If the editor is
requested from the editorManager with the updated editor id, the old
editor ( with the old version node ) is returned and closed.  The new
editor ( with the new version node ) is still visible.  See cms (
org.hippoecm.frontend.editor.AbstractCmsEditor#start )

Also in that case the editor with the variant is not updated.

This fix fixes both problems.