CMS-11269 Move "picker" resize code to separate class / use auto-height
To simplify the impl I moved the resize code into a separate class.
The resize code path for a dialog containing a picker will only be used
if the dialog is resizable.

We can optimize the element lookup a bit further to check if the element
is visible in the DOM by leveraging document.body.contains(el) if
required.

The element lookup now starts with the "hippo-picker" element, and from
there tries to find child elements by class-name.

I removed the restoreDatatableHeight() call as it seems to no longer
add value, but I might have missed something.

I also set the LinkPickerDialog to height=AUTO again, as the
image-picker dialog would now always render a vertical scrollbar; it
adds an upload panel which makes it not fit. In order to work around
not having a minimal height, I added code that calculates the current
height.