CHANNELMGR-2148 Override and return to persisted overlay state
When a channel project is selected which is not allowed to be edited the
overlay and overlay toggles should be disabled. When switching to
another branch or master after that the persisted state in local storage
should again be used.

(cherry picked from commit 11aec94a246eadb5f4ce80ebc6f30e07074890b8)