HSTTWO-4467 Support per hst configuration whether user can add channels
Before, there was a single check 'canUserModifyChannels' which would
return true or false. The check was to indicate whether the current
cms user was allowed to add channels or not if there were existing
blueprints. Since MSDTS setup (multiple hst configurations), this wasn't
sufficient any more: If can be the case that a webmaster is allowed to
add channels in, say, /hst:site but not in, say, /hst:intranet.

To fix this, I dropped 'canUserModifyChannels' completely, and instead
only return blueprints for which the current cms user is allowed to
create channels