CHANNELMGR-2159 Drop keep alive pinger
The keep alive pinger was meant to keep website webapp http session
alive because that contained valuable information like which channel
was rendered. However, all this knowledge is now tied to the cms http
session. When a site webapp http session expires, the SSO connection
is just done again and everything should work again. Not need for a
keep alive pinger that keeps the website http sessions alive