HSTTWO-4472 Move HstSiteMapItemHandler creation and registry to platform
The HstSiteMapItemHandler creation and registry has now moved to platform
AND, the HstSiteMapItemHandler instances are now tied to the VirtualHosts
object. If the VirtualHosts is garbaged, automatically, the
HstSiteMapItemHandler instances will be garbaged as well (they are
referenced by weak references in the registry for this reason)

Downside is that #destroy won't be invoked any more (I did not want to
hook into the finalize method). However, I am pretty sure #destroy has
never done anything for any known HstSiteMapItemHandler! So just deprecated
it. If it is currently used by an end project, they have to rewrite their
logic which shouldn't be hard