HSTTWO-4379 Tie the component registry to virtualhosts lifecycle
The VirtualHost object now has a component registry instance. When
the VirtualHost is invalidated, we clear the component registry of that
instance. However, there might still be some requests running for that
VirtualHost instance which can register new components. Hence, in the
CleanupValve we clear the component registry again if the component
registry is awaiting termination (where termination is actually being
GC-ed)