CHANNELMGR-2176 Let UiExtension.register() return a promise
The promise is resolved with the 'ui' object when everything goes well.

The promise is rejected with an 'error' object when something goes
wrong. The 'error' object has two properties: 'code' (which can be
used by code to understand the error case) and 'message' (a
human-readable explanation of the error).

The errors generated by Penpal are converted to our own error format,
since we don't want to expose any Penpal internals / error codes.
This gives us the freedom to change the underlying library without
changing the API if we want.

The promise constructor can be changed via configuration passed to
the 'register' call.