CHANNELMGR-2154 Move syncPreview() to editComponentMainCtrl
SyncPreview should only be called when a component is being edited,
so now it registers on the events 'componentMoved' and 'createdOverlay'
and unbinds to these events when it is destroyed.
This makes EditComponentService.syncPreview also easier since we know
for sure that a component is being rendered when it is called; no need
to check the state.

Also moved HippoIframeService.signalageLoadCompleted() back to the
RenderingService so it will be called when SpaService.API.createOverlay
is called.