CHANNELMGR-2188 Re-enable interception of external links
In preview mode the HST will render all internal links as paths, even
the fully qualified ones. So any link that starts with a scheme is an
external one.

Reading a link's href value is a little tricky: reading it as a property
will make the browser resolve relative links, so you always get a fully
qualified link. Which does not work if we want to recognize paths...

Yet reading the href value as an attribute gives the original source
code value, which can be used to recognize paths. That also matches
nicely with links in SVGs. Instead of the complex instanceof logic
(which turned out not to work anymore anyway, at least not in Chrome 70)
we can simply read the attribute 'xlink:href' to get the href value.