CHANNELMGR-2202 Add Bash script to release NPM package
Releasing will be a manual step since you need to login to NPM first,
which (possibly) triggers a one-time-password if you have 2-factor
authentication enabled for publishing. The version to release can be
passed to the script. Since the version passed to the script will be
leading, the default version in package.json is set to 0.0.0.