CHANNELMGR-2164 Reload project(s) when switching perspectives
Moving the check if project is the same or not to the actual callbacks
being run when the selected project is updated allows to be more
specific on when to execute an action.
Using a getter for the toggle will update the toggle automatically on
the next digest cycle when projectservice projects change.