CHANNELMGR-2204 Reload project(s) when switching perspectives
Moving the check if project is the same or not to the actual callbacks
being run when the selected project is updated allows to be more
specific on when to execute an action.
Using a getter for the toggle will update the toggle automatically on
the next digest cycle when projectservice projects change.

(cherry picked from commit 29739dc7db2ed6ad05b45a344a8924edea63f52f)