CHANNELMGR-2206 Check node type of request
There are two types of requests:
- requests created by authors, of type hippostdpubwf:request
- requests created by editors, of type hipposched:workflowjob
Any other kind of node type for a request must be considered an error,
so it is logged as a warning.

Because the request types have different properties they must be treated
separately.

Please note that the returned error reason currently always is
"The document is waiting to be published.", which is not always correct.
Fixing that issue will be done via another JIRA issue.