CMS-11580 Revert "CMS-11504 Remove state icon from history dialog"
This change broke functional tests. That is not the only reason to
revert the change. Another one is that the revision history as a
feature is already quite confusing to customers. They mistakenly view
the revision history as an audit log, while it is actually a list of
historic versions. Keeping the changes to a minimum is therefore best.

This reverts commit 3b99349