CHANNELMGR-2140 Hide 'defined in' blocks in TypeDoc documentation
The 'defined in' block mention the source file name and line number at
which a documented item is defined. This information is not relevant
for end users of the library, so it's better left out to make the
documentation more clean.

Unfortunately there's no official option to skip generating the
'defined in' blocks. We therefore patch the generated CSS instead.
Using 'minicat' ensures the build also works on Windows.