CHANNELMGR-2226 Use branchId of document for updating a document
If a user creates new content while he selected a project he should get
the option to add it to the project or keep it on core. So, when the
user saves the document at the end op step two the "Add to project"
dialog must be shown. So instead of looking at the selected branchId
present in the CMS session context the server should use the branchId
of the document to perform the document update actions. If a new document
is created the server returns a JSON document with branchId master. So
when the client code then calls updateDocument it will do so with
branchId maste. In the response the branchId is then still master. The
client code will then detect that the document does not belong to the
project yet and will automatically show the dialog.

(cherry picked from 9a5f1298)