CMS-11583 Move unsupported-browser warning into separate label
The unsupported browser warning should always be visible, instead
of only once. This requires it to be moved into a separate label,
instead of a message in the feedback panel (these are only rendered
once).

This simplifies the BrowserCheckBehavior, as it now only sets the host
component visibility, instead of logging a warning.

To simplify the constructors, the LoginConfig is introduced.