REPO-2167 Add branch awareness to derived data engine
Define the modified variant as the ancestor of the modified node or
the modfied node itself if it exists.

Define the accessed variant as the ancestor of the accessed property if
it exists.

For example if D is the name of a document and M is the modified node
then for .../D/D/M .../D/D is the variant. It could be that the modified
node is not a variant or a descendant of a variant.

If for both the accessed property and the modified node variants are
found and those variants both have a branchInfo mixin and their branch id's do
not match then the accessed property is read from version history.

If it cannot be found a RepositoryException is thrown.

If the accessed variant has a branch id and the modified variant doesn't
or vice versa the a DerivedDataConfigurationException is thrown. That
situation can lead to unexpected results. If you define derived data
where the accessed and the derived properties are both descendants of
the same handle that situation cannot occur.

Only in very peculiar configurations the situation could occur.