CMS-11583 Disable submit button & set form fields to readonly on submit
Disabling a form field prevents it from being submitted in the form POST
so instead set form fields to readonly by using the :input selector from
jquery which selects all form fields.

A readonly field can still be focused, so use a 'form-disabled' class on
the form to remove the border from a focused field. Also adds opacity to
input/select elements.

A select widget can not be set to read-only so the best I could find is
to disable all option elements apart from the one that was selected.
This prevents a user from selecting a different option.
The other
workaround I found is to add a hidden input with the same name as the
select widget, which contains the value of the selected option, but this
is not a one-liner and seems a bit more of a workaround.