CMS-11605 Calculate tree width on Edge
Added browser-specific workaround for Edge; ensure the tree items can
dictate the width of the hippo-tree element. This is done by CSS on
Chrome, Firefox and Safari but Edge does not support it. Previously we
only used script to calculate the appropriate width on IE but the UA
check is different for Edge so that never fully worked until now.

I've bumped the magic width to ensure the blue-marker wraps the full
tree item.