CHANNELMGR-2232 Remove 'site is down' message
The 'site is down' message used to be shown when the site webapp was not
running, in which case the channels could not be retrieved. In version
13 the channels are retrieved from the platform, so the site does not
need to be up to know there are no channels.

Due to a bad merge the existing logic was also broken, since the
rootPanel was still deferenced when it was null.

Removing the whole 'site is up' check completely seemed a cleaner
solution.