CMS-11121 Do not redraw editor if node.isCheckedOut() did not change
The console node editor can be displayed in read-only mode for nodes
that are checked-in. Previously, any change to the nodeModel would
trigger a redraw of the full editor, resulting in focus loss.

With the current change, the editor is redrawn when the nodeModel
changes (user select a different node) or when a change is made to the
node that changes it's checked-out/checked-in state.