CHANNELMGR-2241 make sure exceptions are caught and logged as warning
Logging only IllegalStateException on info level is problematic:

First of all, end projects can hook in custom channel filters which can
throw *any*  runtime exception which is not caught unless it is an
IllegalStateException. Such exceptions then bubble all the way up and
are handled by wicket, resulting in long stacktraces and a channel mngr
dashboard which does not respond when clicking on it. And equally bad,
when an IllegalStateException you get an empty channel mgr overview but
nothing will be logged (unless ChannelStore is at info log level which
is unlikely and developers can't know)

Better to always log a warning and catch all exceptions