CMS-11624 Use name attribute to find the timezone select widget
Previously, we used the ID attribute to find the select dropdown, but
in the new login panel, the ID of the select is generated by Wicket.

For simplicity, we use the (existing) name attribute to find the select
element. The hard way would be to pass the wicket ID to javascript.