CMS-11629 Only if JcrSessionModel is pressent there is a real user session
When the HstRequestContext session gets set in Main.java for every
request, it also frequently gets the anonymous jcr session...most likely
for wicket resources or something. This is undesirable since it results
in very many logins of the anonymous session.

Instead, only set the real user session, which is there when there is
a non-null JcrSessionModel