CHANNELMGR-2250 Make sure repository part does not pull in cms code
Since the [platform + site] setup should not pull in any wicket code,
and since the content-service module has a dependency on wicket, the
hippo-addon-channel-manager-repository module should not pull in
hippo-addon-channel-manager-content-service. Since the latter contained
yaml configuration, I have moved the yaml config to the repository
module, and move the dependency on the content-service in the repo
module to frontend-ng.