CHANNELMGR-2260 Don't get antiCache from top window
If the antiCache field is available in the HippoGlobal object then it
can simply be used as is. There is no need to get it from the top
window. The top window might in the future be some other enclosing
parent iframe without the Hippo object in it, so then the Hippo
object might even not be present in the top window.