HSTTWO-4533 Support for custom website HstSiteProvider(s)
Per HST site webapp, it should be possible to plug in a different custom
website HstSiteProvider. Since the HstSiteProvider gets invoked from the
platform webapp code, we provide a cross webapp shared
CustomWebsiteHstSiteProviderService which can be used by end projects
to inject a custom website HstSiteProvider. A custom website
HstSiteProvider will be invoked by the classloader of the HST site
webapp that injects the custom HstSiteProvider