HSTTWO-4533 Avoid easy to trigger recursion
If a downstream project or end projects implements an HstSiteProvider and
would for example write something like this:

HstSite getHstSite(CompositeHstSite hstSite, HstRequestContext requestContext) {

    hstSite.getName();
}

then recursion would happen and a stack overflow. Reason is that many of
the HstSite method invocation on a CompositeHstSite result in
getActiveSite which triggers again the code above.

Instead, we should only pass in the non composite master and
branches to avoid easily triggered recursion