CMS-11210 Add "infoRequest" request type
In org.hippoecm.frontend.plugins.reviewedactions.RequestsView#onPopulate
the requests from the hints are used to populate request items that are
shown in the editor.

In case the user only belongs to the author group not all request were
added. That is fixed by adding a "infoRequest" hint. As a side effect
the requests are added. See hippo-repository related commit.

Make sure this hint is actually used by the cms, if only to prevent that
the action is deleted from the scxml because it's not used. An
additional reason is that changes in the value of the "infoRequest" hint
result in actually hiding or showing the info item.

Remove the isVisible methods and replace them by explicitely setting the
visibility after adding the item for performance reasons. isVisible is
used every time wicket redraws its page.