CMS-11210 Always add all requests
Always add a "infoRequest" so requests are always added to the feedback
and eventually to the hints. Right now the infoRequest is always added
and is true.

As a side effect of the hippo:requestAction the "requests" object in
scxml is added to the feedback. See
org/onehippo/repository/documentworkflow/action/RequestActionAction.java:96

If not a single requestAction is added the "requests" object is not
added to the feedback. As a consequence the frontend code never is able
to display any representation of the requests.

The requests were only added if the request is a "workflow request" and
the current user belongs to the editor group. Requests can be workflow
requests, like "request publication" or scheduled request, like
"schedule publication". All these requests should be visible to
authors.