HSTTWO-4540 Support runtime virtual host injection in the model
When a requested host does not exist, we add it on the fly, by default
mimicking the dev-localhost setup.

A new hostgroup is added to all hst configs. The host group name is a
random uuid and only used internally.

Whenever the HST model invalidates, all known runtime hosts are rebuild.

This is very cheap since they a thin wrapper classes