CHANNELMGR-2251 show the validation messages for channel setting fields
The messages were not correctly passed to the property field component. When they showed up some reordering of the html elements was necessary to place the error correctly.
Rephrased the validation text along Angular Material guidelines: give the user a hint what to do to fix the problem instead of describing it.