HSTTWO-4541 HstSiteMapMatcher for correct webapp should be used
Since every HST webapp can have its own unique HstSiteMapMatcher, we
must ensure the correct one is returned. If we use
HstServices.getComponentManager we get the component mgr for the webapp
of the class loader of the object that contains
HstServices.getComponentManager : For the MountService, this will return
the component mngr for the platform, since the platform webapp creates
MountService instances. Hence getting the HstSiteMapMatcher in this way
always returns the platform HstSiteMapMatcher which is incorrect. Instead,
the HstSiteMapMatcher should be returned for the webapp for which the
MountService instance is created.