CMS-11667 Replace CMS built-in validators with validation-service module
The validator-service-plugin will now use the new module validators if
it can not locate the built-in validators of the CMS, otherwise the old
behavior remains.
We still need to write an updater script to remove the old validators
if there is no custom configuration found.

To use the new validators in your dev environment for now we need to
manually delete them from the repository at
/hippo:configuration/hippo:frontend/cms/cms-validators