REPO-2180 [Backport 12.6] Instead of storing hashed pwd, store a secret
The secret is tied to the Credentials instance. If there is a secret
attribute present, the decrypted value must be the same as the
Credentials system identity. If that is the case, the authentication
for that Credentials instance has already been done

(cherry picked from commit be8e72176856e38e54f864308d6f470feb41c26d)