CMS-11752 Copy response headers
Copy response headers, because chrome needs a mime type to display
svg files as images.

When upgrading the angular version 7 the svg
images are not served anymore as data url. See WPM-1168.

The code is loosely inspired by:
https://github.com/mitre/HTTP-Proxy-Servlet

This module is used by hippo-addon-targeting. The version is the
same as defined in its dependency management section.