CMS-11668 Handle "required" validation, pass correct value to validators
The old 'validateRequired' code actually did the logic of the 'nonempty'
validator, and has been removed again. The "required" validation is now
only concerned with checking whether there is at least one value. It's
done in the "validate()" method before any other validator, instead of
being part of the cardinality check before writing the actual value.
It does not return a message since the frontend is expected to perform
required validation itself.

Because 'validateRequired' was not needed anymore in subclasses, the
method hierachy changed a bit. The 'validateValue' method now validates
a single value of a field. Only 'validateValue' needs to be overridden
by CompoundFieldType and ChoiceFieldType to get recursive validation of
their children.

Each validator should get a field value specific for the field type it's
configured on. For example:
 - a Date field should pass a Calendar to its validators
 - an imagelink field should pass a Node to its validators
 - etc.
Each primitive field type therefore now return the value to validate by
overriding getValidatedValue().