CMS-11752 Append values multivalued header
Support headers with multivalued header like Content-Language.

Add cookie headers and location to ignored headers, because the Location
headers should not be copied by a proxy and to proxy cookies from the
response does not seem to make much sense for the intented use of the
Filter, while it add a lot of complexity.

Add status code to response.