CHANNELMGR-2305 Validate document fields when auto-saved
When a field is auto-saved to the draft node, the backend now also
validates the field value(s). Any validation errors result in errorInfo
objects added to the invalid field values, which are returned to the
frontend. The frontend adds these errorInfo objects to the edited field
values so a validation message appears.

Some tricky details:

- In case of multiple field values and multiple errors, the validation
  error first invalid value is shown.

- The ink ripple on the save button has been disabled, otherwise a
  validation result that marks the form invalid (and hence disables the
  button) stops any animation on the button. When the save button is
  hovered, that results in a frozen hover state, which looks very odd.
  To reproduce the original issue, put back the ink animation, make a
  field invalid and click 'Save'.

- Only on save the first invalid field is focused, otherwise per-field
  validation will also focus a field once it becomes invalid.