CMS-11763 Skip "required" validator & don't keep reference to Validator
The "required" validator is built-in, meaning that it is not implemented
as a separate validator but as part of the generic #validate() call.
Previously, missing validators would simply be ignored, but now we log
a warning when it is not found (not in CMS and not in
modules/validation)

Apart from that, we should not keep a reference to a Validator as that
circumvents users from changing the configuration.