HSTTWO-4540 Extra assertions that the template host group is correct
A host group can only be used if it has a single host with directly an
hst:root mount below it. All other cases, the host group cannot be used
as a template hostgroup for runtime mounts.

Also when the autohosttemplate property is present on multiple host groups,
we need to log an error since this is not supported