HSTTWO-4556 Fix rendering incorrect branch from document editor menu
This issue turns out to have many angles and hard to get / keep it right.
We need to account for the following use case:

1) When the Channel Mngr ChannelStore loads the channels by the /cms
webapp via

ChannelServiceImpl#getPreviewChannels(Session, String)

then the HST request is not marked as a cms request
(HstRequestContext#isCmsRequest() is false), which is in fact correct:
It is confusing since the name isCmsRequest() can better be
#isChannelMngrPreviewRequest() because that is what is really meant
with this 'isCmsRequest' : We need to modify this getter, I'll create a
separate issue for this. Since the #isCmsRequest returns false, the
DelegatingHstSiteProvider will use the websiteHstSiteProvider for
returning the right HstSite, but that is clearly wrong! Also the
channelManagerHstSiteProvider should not really be needed, since the
ChannelStore should just always load the MASTER preview channels!

2) when the

DocumentServiceImpl#getPreviewChannels(Session, String, String, String)

is invoked from the CMS menu in the document editor, the HST request is
not a cms request (ok, it is a platfor request but not a channel manager
preview request), and as a result, the same problem as with (1) occurs
that the websiteHstSiteProvider is used returning the wrong HstSite

Also there is another problem: When for a single request

DelegatingHstSiteProvider#getHstSite

is invoked multiple times, the last call sets the 'RENDER_BRANCH_ID'
requestcontext attribute as string...this is wrong, since per HstSite
the branch or the master (if branch misses) can be used. Instead,
the RENDER_BRANCH_ID attribute should store a map