HSTTWO-4557 deprecate confusing method and replace with correct name
A lot of code in the HST rely on whether the request is a cms request
or not.

Using however HstRequestContext#isCmsRequest is however very confusing
since version 13.0.0 : For example all HST requests via the platform
webapp (/cms) are NOT cms requests! They are just plain HST requests
that happen to go over the /cms webapp. What 'isCmsRequest' really
means is that an HST Site webapp is hit (for example /site) OVER
the host of the CMS : Aka, it is a 'channel manager preview request'!!
A channel mgr preview request must for example decorate Mounts to the
preview mount. A request over the cms webapp, for example /cms/_rp
never needs decorated mounts. Hence, we should deprecate #isCmsRequest
and give it a proper name since it is so confusing now and results in
bugs because the behavior is different than people expect from the name

Before 13.0.0, there was no confusion since the /cms webapp did
not contain the HST