HSTTWO-4557 /cms/_rp calls are platform requests, not channel mgr preview
Apparently however, before this was added because of downstream projects
relying on it. I cannot easily find which downstream project would rely
on it. Relevance perhaps? Either way, since I don't see any reason to
keep it and since it seems wrong, I want to get rid of it. We need to
carefully validate that nothing breaks after this change, certainly
in relevance, for example the persona html comments