CMS-11757 Redirect to the login page when a spurious pinger hits the CMS
A browser tab with a CMS that has been restarted will still send pinger
requests. These requests will hit the login page, which for this purpose
already contained a fake pinger in the form of a Label with the same
Wicket ID "pinger" so its Wicket path matched that of the pinger in the
CMS.

Such a pinger request is an Ajax request for an expired page. The
behavior regarding such requests changed subtly in Wicket 7 (see
ListenerInterfaceRequestHandler#respond). Wicket 6 used to redirect such
requests to the current page (i.e. the login page), so a spurious pinger
call would effectively show the login page again. Wicket 7, on the other
hand, added a check whether the targeted component is stateless or not.
Since the login page added a Label, and a Label is stateless, the
redirect no longer happened and an exception was logged instead.

Fixed by always using an actual Pinger component and making that
stateless. To make the intent more clear I've deprecated the existing
constructors and introduced static builder methods instead (the
constructor with only the ID was not used at all).