HIPPLUG-1583 Reload tree and restore state when moving a node
It is currently not possible to move a node and keep the whole tree in
sync, without a reload of the whole tree. To ensure the user experience
does not deteriorate, we try to restore the state with regards to
previously selected and expanded nodes after the reload is done.