HSTTWO-4552 Create BeanBuilderService for bean generation
Create a new AbstractBeanBuilderService to define bean creation
logic to use in BeanWriter and DynamicBeanService. Simplify the
bytebuddy implementation in DynamicBeanService to cover
all document types and compound types. Listen the requests which
are made to the HstRequestContextImpl to verify whether there is
any version change in ContentTypeService to invalidate runtime
generated dynamic beans and regenerate them again.