CMS-11745 Move initial height in pixels to extension config
The string config property 'initial.height.in.pixels' per OpenUiString
field is replaced with a long config property 'initialHeightInPixels'
per UI extension.

The long is cast to an int since that's big enough to hold a number of
pixels. Also, a Java int still fits in a JavaScript number while a Java
long does not.

An initial height per extension makes more sense than an initial height
per field: the optimal initial height depends on the extension and its
configuration, which are determined by the extension's URL and its
'config' property, respectively. So better keep the initial height next
to those properties.

The initial height does not need to be validated. An invalid value will
be clamped to the valid range by the frontend code anyway.