CMS-11805 Add auto-save behavior to OpenUiString field
When #setFieldValue is called, a save in the backend is scheduled with a
delay of two seconds. If the plugin is destroyed while a scheduledSave
is still pending, it will be ignored.

To prevent data-loss within the two second window, we need a blur event
to be triggered by the extension.