CMS-11787 Refactor generic UI extension logic into separate behavior
The OpenUiBehavior can be added to any Wicket component, and will let
Penpal create an iframe element inside it that loads a UI extension.

The Wicket component should have a .js sibling file that defines a
JavaScript class. This class must define life cycle hook methods for
the Penpal connection (onConnect and onDestroy), and a getMethods() call
that returns the Penpal parent methods. Any parameters defined in the
getJavaScriptParameters() method of the Open UI plugin will be passed to
the constructor of the JavaScript class.

As a result, the OpenUiStringPlugin now only defines the logic specific
to that plugin. All generic logic (i.e. creating/destroying the Penpal
connection and the getProperties() method) is handled by the
OpenUiBehavior.