HSTTWO-4577 Use HippoEssentialsGenerated instead of AutoEnhancedBean
Run dynamic bean generation for the existing beans according to
HippoEssentialsGenerated annotation.
Not generate a new class if the existing bean is up-to-date
Add a system property to be able to enable/disable dynamic bean
functionality.
Improve the bean generation logic of dynamic object converter.
Add integration test for compound type inheritance.
Fixed of generating methods that are exist in parent class.