HSTTWO-4580 Drop usage of hst:defaulthostname
The property hst:defaulthostname was quite crappy since the beginning
and in general should never be used for years already. However with the
introduction of branch support and with runtime HST hosts support, the
property really breaks features. Hence, drop its support now formally
though it was already unofficially a do-not-use property