HSTTWO-4560 Make sure page model api mount also has Channel object
The page model api mounts should also return a channel object. To avoid
the channel object with the same id to occur twice in the channel map
of the virtualhosts service, I skip the channel objects for page model
api mounts. Perhaps in the future we should introduce a clearer check,
like isCanonical(). For now, this suffices.