CMS-11787 Add new Dialog#onCancelFromCloseButton hook
Clicking the close link (X) of a dialog should set parameter
AbstractDialog.cancelled to true in order to execute the correct
ScriptAction in org.hippoecm.frontend.dialog.Dialog#onClose.

This is required for the openui dialog as it should reject a promise
when the dialog is cancelled.

Overall, this improves dialogs that require specific behavior on cancel,
as it would previously not be invoked when clicking the (X). Now it is
similar to pressing escape or clicking the 'cancel' button in the bottom
row.