CMS-11557 Show all revisions
Show all revisions, so that the previous functionality is restored.
This is now possible, because the correct menu ( workflow menu or
revision menu ) is shown ( See WPM-1193 ).

Disable revision if its frozen node:
is the same node as the frozen node displayed in the editor
or
has the last changes made to the
unpublished variant displayed in the editor

Disable that revision to prevent the same frozen node to be opened in a
new editor tab. That would result in two tabs displaying the same branch
of a document and both displaying the workflow menu.

Also disable that revision if the unpublished variant is shown, to be
consistent with the previous case ( because from a functional
perspective there is no difference ) and because it's technically less
work ( we only have to check the label of the revision ).

Add the "disabled" css class ( and relevant styling ) and disable the
functionality to open a new tab, to prevent a new tab from being opened
and provide user feedback.

Pass on the branchId from the Plugins all the way down to the History
Dialog so the relevant Revision can be disabled.